split commit.go
